+++ rancher2/resource_rancher2_cluster.go
@@ -234,7 +234,7 @@ func resourceRancher2ClusterUpdate(d *schema.ResourceData, meta interface{}) err
 	}
 
 	stateConf := &resource.StateChangeConf{
-		Pending:    []string{"active", "provisioning", "pending", "updating"},
+		Pending:    []string{"active", "provisioning", "pending", "updating", "upgrading"},
 		Target:     []string{"active", "provisioning", "pending"},
 		Refresh:    clusterStateRefreshFunc(client, newCluster.ID),
 		Timeout:    d.Timeout(schema.TimeoutUpdate),