+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.8.2 (Unreleased)
+## 1.8.2 (April 02, 2020)
 
 FEATURES: